convert types, renames

+++ internal/currency/usd.go
@@ -9,21 +9,21 @@ import (
 
 // USD describes USD currency.
 type USD struct {
-	cents int64
+	cents int
 }
 
 // Dollars converts dollars to USD amount.
-func Dollars(dollars int64) USD {
+func Dollars(dollars int) USD {
 	return USD{dollars * 100}
 }
 
 // Cents converts cents to USD amount.
-func Cents(cents int64) USD {
+func Cents(cents int) USD {
 	return USD{cents}
 }
 
 // Cents returns amount in cents.
-func (usd USD) Cents() int64 { return usd.cents }
+func (usd USD) Cents() int { return int(usd.cents) }
 
 // String returns the value in dollars.
 func (usd USD) String() string {
